change float to int in range

+++ person-recognition/videos_tracker/centroidtracker.py
@@ -63,7 +63,7 @@ class CentroidTracker:
 	def __get_largest_faces_args(self, face, k=5):
 		faces_size = [abs(bottom-top) * abs(right-left) for (left, top, right, bottom) in face.rects]
 		sorted_face_args = np.argsort(faces_size)
-		largeset_face_args = [sorted_face_args[i] for i in range(len(sorted_face_args)/2, len(sorted_face_args)/2 + k) ]
+		largeset_face_args = [sorted_face_args[i] for i in range(int(len(sorted_face_args)/2), int(len(sorted_face_args)/2 + k)) ]
 		return largeset_face_args
 
 	def __create_person_encodings(self):
